backend/requirements.txt
# FastAPI 및 웹 프레임워크
fastapi==0.112.0
uvicorn[standard]==0.24.0
pydantic==2.7.0
pydantic-settings==2.1.0
anyio==4.5.1         # FastAPI와 호환되는 버전으로 명시

# 데이터베이스
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# MQTT
paho-mqtt==1.6.1

# Kafka 스트리밍 플랫폼
kafka-python==2.0.2
confluent-kafka==2.3.0
aiokafka==0.12.0

# ML/DL
scikit-learn==1.3.2
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
pandas==2.0.3
scipy==1.10.1
xgboost==2.0.2
joblib==1.3.2

# 모델 서빙
bentoml==1.4.18

# 시계열 데이터 처리
tslearn==0.6.2
statsmodels==0.14.0
plotly==5.17.0

# 인증 및 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# 알림 시스템
requests==2.31.0

# 모니터링
prometheus-client==0.19.0

# 유틸리티 및 테스트
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
redis==5.0.1
celery==5.3.4
schedule==1.2.0

# 로깅 및 추적
loguru==0.7.2
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0

# 개발 도구
black==23.11.0
isort==5.12.0
flake8==6.1.0
